moved integration tests
